% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\docType{data}
\name{gene_bridge_data}
\alias{gene_bridge_data}
\title{A pre-processed dataset for the GeneBridge package}
\format{
A data frame containing orthologous groups (OGs).
}
\source{
This package.
}
\usage{
data(gene_bridge_data)
}
\value{
A pre-processed dataset.
}
\description{
Dataset used to demonstrate GeneBridge main functions.
}
\details{
The list with 4 R objects to be used for demonstration purposes 
in the GeneBridge vignette.

\describe{

\item{ogdata}{
A data frame with three columns listing orthology 
annotation retrieved from the STRING database (http://string-db.org/), 
release 9.1. Column 1 = Ensembl protein ID; column 2 = NCBI species ID; 
column 3 = OG ID. Note: This dataset is to be used for demonstration 
purposes only as it represents a subset of the STRING database; in order 
to reduce the dataset size, orthology annotation was mapped to genome 
stability genes (Castro et al.).
}
\item{sspids}{
A data frame containing the species listed in STRING database 
(http://string-db.org/), release 9.1. Column 1 = NCBI species ID; 
column 2 = NCBI species name;column 3 = species domain (eukaryotes).
}
\item{ogids}{
A vector listing IDs of the OGs available in the 'ogdata' object.
}
\item{phyloTree}{
An phylo-class object for the eukaryotes listed in the STRING database, 
release 9.1.
}
}
}
\examples{
data(gene_bridge_data)
}
\references{
Franceschini et al. STRING v9.1: protein-protein interaction networks, 
with increased coverage and integration. Nucleic Acids Research, 
41(Database issue):D808-15, 2013. doi:10.1093/nar/gks1094.
}
\keyword{gene_bridge_data}
