% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\docType{package}
\name{GeneBridge-package}
\alias{GeneBridge}
\alias{GeneBridge-package}
\title{GeneBridge: large-scale evolutionary analysis of orthologous groups}
\description{
GeneBridge is designed to assess the distribution of orthologous groups in 
a given species tree. It implements the Bridge algorithm to determine the 
evolutionary root of genes in large-scale evolutionary analysis.
}
\details{
\tabular{ll}{
Package: \tab GeneBridge\cr
Type: \tab Software\cr
License: \tab GPL-3\cr
Maintainer: \tab Mauro Castro \email{mauro.a.castro@gmail.com}\cr
}
}
\section{Index}{

\tabular{ll}{
\link{newBridge}:
\tab Constructor of Bridge-class objects.\cr
\link{runBridge}:
\tab Run gene root inference with the Bridge algorithm.\cr
\link{getBridge}:
\tab Accessors for fetching slots from a Bridge object.\cr
}
Further information is available in the vignettes by typing
\code{vignette('GeneBridge')}. Documented topics are also available in
HTML by typing \code{help.start()} and selecting the GeneBridge package
from the menu.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sysbiolab/GeneBridge}
}

}
\author{
\strong{Maintainer}: Mauro Castro \email{mauro.a.castro@gmail.com} (\href{https://orcid.org/0000-0003-4942-8131}{ORCID})

Other contributors:
\itemize{
  \item Leonardo Campos [contributor]
  \item Rodrigo Dalmolin [contributor]
}

}
\keyword{internal}
