% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{newBridge}
\alias{newBridge}
\title{Constructor of Bridge-class objects}
\usage{
newBridge(cogdata, phyloTree, spid, cogids = NULL, verbose = TRUE)
}
\arguments{
\item{cogdata}{A data frame with COG data.}

\item{phyloTree}{Object of class \code{phylo}.}

\item{spid}{A single string or integer specifying the reference species 
to be used in the Bridge algorithm. This species should be listed in the 
\code{phyloTree}.}

\item{cogids}{An optional data frame with COG annotation. Alternatively, 
this can be a character vector with COG IDs.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when \code{verbose=TRUE}) or not (when \code{verbose=FALSE}).}
}
\value{
A preprocessed \linkS4class{Bridge} class object.
}
\description{
\code{Bridge} is a constructor of Bridge-class objects.
}
\examples{
# Load datasets used for demonstration
data(cogData)

# Create an object of class 'Bridge' for H. sapiens
gbr <- newBridge(cogdata, phyloTree, spid="9606", cogids=cogids)

}
\seealso{
\code{\link{runBridge}}
}
\author{
Sysbiolab.
}
