% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlots.R
\docType{methods}
\name{plotBridge,Bridge-method}
\alias{plotBridge,Bridge-method}
\alias{plotBridge}
\title{plotBridge}
\usage{
\S4method{plotBridge}{Bridge}(
  gbr,
  whichOG,
  fname = "gproot",
  width = 4.5,
  height = 6.5,
  cex.lab = 0.3,
  cex.nodes = 0.6,
  adj.tips = c(1, 0.5),
  lab.offset = 1.5,
  col.tips = c("green2", "grey"),
  col.edges = c("black", "grey"),
  col.root = "red",
  plot.sspnames = TRUE,
  plot.subtree = FALSE,
  plot.lcas = FALSE
)
}
\arguments{
\item{gbr}{A preprocessed \linkS4class{Bridge} class object.}

\item{whichOG}{A single string indicating the OG to be plotted.}

\item{fname}{A name for the file in which the plot will be generated.}

\item{width}{A single number specifying plot 'width' in inches.}

\item{height}{A single number specifying plot 'height' in inches.}

\item{cex.lab}{A single number specifying expansion factor for tip labels.}

\item{cex.nodes}{A single number specifying expansion factor for 
node symbols.}

\item{adj.tips}{A vector with two numbers.}

\item{lab.offset}{A single number to offset labels.}

\item{col.tips}{A vector (length=2) specifying colors of the tips.}

\item{col.edges}{A vector (length=2) specifying colors of the edges.}

\item{col.root}{A single color specifying the color of the inferred root.}

\item{plot.sspnames}{A single logical value specifying whether 'ssp' names 
should be plotted.}

\item{plot.subtree}{A single logical value specifying whether a 
sub-tree (that includes the root) should be plotted.}

\item{plot.lcas}{A single logical value specifying whether a species tree 
should be generated mapping the positions of all possible roots.}
}
\value{
A pdf file.
}
\description{
The \code{plotBridge} generated a tree diagram for the inferred 
evolutionary root of a given OG.
}
\examples{
# Load datasets used for demonstration
data(cogData)

# Create an object of class 'Bridge' for H. sapiens
gbr <- newBridge(cogdata, phyloTree, spid="9606", cogids=cogids)

# Run the Bridge algorithm
gbr <- runBridge(gbr, nPermutations=100)

# This example plots 'NOG40170' in the phyloTree
plotBridge(gbr, whichOG="NOG40170")

}
