% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPlots.R
\docType{methods}
\name{plotBridgeSimulation,Bridge-method}
\alias{plotBridgeSimulation,Bridge-method}
\alias{plotBridgeSimulation}
\title{plotBridgeSimulation}
\usage{
\S4method{plotBridgeSimulation}{Bridge}(gbr)
}
\arguments{
\item{gbr}{A preprocessed \linkS4class{Bridge} class object.}
}
\value{
A ggplot object.
}
\description{
The \code{plotBridgeSimulation} generated a ggplot2 object for 
statistics inferred by the Bridge algorithm.
}
\examples{
# Make a 'phylo' tree
tree <- symmetricBranching(n.branches = 10)

# Simulated 100 ROGs
gbr <- simulateRogs(phyloTree = tree, n.rogs = 100)

# Run the Bridge algorithm
gbr <- runBridge(gbr)

# Check predictions
plotBridgeSimulation(gbr)

}
