% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{runPermutation,Bridge-method}
\alias{runPermutation,Bridge-method}
\alias{runPermutation}
\title{Check consistency of root placement}
\usage{
\S4method{runPermutation}{Bridge}(
  gbr,
  nPermutations = 1000,
  pAdjustMethod = "bonferroni",
  verbose = TRUE
)
}
\arguments{
\item{gbr}{A preprocessed \linkS4class{Bridge} class object.}

\item{nPermutations}{A single integer (>1) specifying the number of
permutations used to compute a null distribution for the inferred roots 
in the species tree.}

\item{pAdjustMethod}{A single string specifying the p-value 
adjustment method to be used (see 'p.adjust' for details).}

\item{verbose}{A single logical value specifying to display detailed 
messages (when \code{verbose=TRUE}) or not (when \code{verbose=FALSE}).}
}
\value{
A preprocessed \linkS4class{Bridge} class object, including 
results from the \code{runPermutation} function.
}
\description{
The \code{runPermutation} function assesses the consistency
of the roots inferred by the Bridge algorithm.
}
\details{
The 'runPermutation' function computes an inconsistency score, 
'Dscore', which is used to assess the significance of the inferred 
root against a null distribution obtained by permutation analysis.
}
\examples{
# Load datasets used for demonstration
data(gene_bridge_data)

# Create an object of class 'Bridge' for H. sapiens
gbr <- newBridge(gene_bridge_data$ogdata, gene_bridge_data$phyloTree, refsp="9606")

# Run the Bridge algorithm
gbr <- runBridge(gbr)

# Assess root consistency
gbr <- runPermutation(gbr, nPermutations=100)
res <- getBridge(gbr, what="results")

\donttest{
# Option: parallel version with the SNOW package
library(snow)
options(cluster=makeCluster(2, "SOCK"))
gbr <- runPermutation(gbr, nPermutations=100)
stopCluster(getOption("cluster")) 
}

}
\seealso{
\code{\link{newBridge}}
}
\author{
Sysbiolab.
}
