% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{simulateRogs}
\alias{simulateRogs}
\title{Simulating Random Orthologous Groups (ROGs)}
\usage{
simulateRogs(
  phyloTree,
  n.rogs = 100,
  gain = 0.2,
  loss = 0.4,
  event.trials = 5,
  interaction.trials = event.trials * 2,
  ref.sp = c("max_space", "top_tip"),
  return.list = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phyloTree}{Object of class \code{phylo}.}

\item{n.rogs}{A single integer (>0) specifying the number of ROGs to be 
simulated.}

\item{gain}{A single numerical value in [0, 1]. Probability of gene gain
events, modeled by the \code{\link{rbinom}} funtion.}

\item{loss}{A single numerical value in [0, 1]. Probability of gene loss 
events, modeled by the \code{\link{rbinom}} funtion.}

\item{event.trials}{A single integer (>0) specifying a fixed number of 
trials associated with each possible outcome (success or failure)  
in modeling gain and loss events. Each trial is independent, so the total
number of random events increases with more trials.}

\item{interaction.trials}{A single integer (>0) specifying a fixed number 
of trials associated with the event order. Each interaction trial 
randomizes the event order, so the event types interact along the 
simulation.}

\item{ref.sp}{A single string specifying which \code{phyloTree} tip will 
be used as the reference species. When ref.sp = 'top_tip', the reference 
species will be the one at the top of the provided \code{phylo} object; and
when ref.sp = 'max_space', the reference species will be the one that 
maximizes the number the possible roots. The reference species will 
establish a "root space", encompassing all ancestors of the reference
species in the tree.}

\item{return.list}{A single logical value specifying to return results
into a \code{list} object. In this case, the list will include all
data objects required to run the \code{\link{newBridge}} constructor.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when \code{verbose=TRUE}) or not (when \code{verbose=FALSE}).}
}
\value{
Either a \code{Bridge} object or a list with data objects required 
to build a \code{Bridge} object.
}
\description{
The \code{simulateRogs} function generates random orthologous 
groups by modeling gain and loss events within a given 'phylo' tree object.
The simulation starts by assigning a random root to each ROG. Following 
that, a vertical heritage is propagated through the tree, extending up to a
designated reference species. The resulting structure is then subjected to 
random gain and loss events. The primary goal of the simulation is to assess
variations from ideal vertical heritage patterns that include the reference 
species.
}
\examples{

# Make a random tree with 'ape'
library(ape)
phyloTree <- rtree(n=25, br=NULL)

# Simulate 10 ROGs
rogs <- simulateRogs(phyloTree, n.rogs = 10, return.list = TRUE)

}
\seealso{
\code{\link{newBridge}}
}
\author{
Sysbiolab.
}
