% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{symmetricBranching}
\alias{symmetricBranching}
\title{Symmetric Branching Tree Generator}
\usage{
symmetricBranching(n.branches = 10, n.tips = seq(1, n.branches))
}
\arguments{
\item{n.branches}{A single integer (>=2) specifying the number of branches
to be simulated.}

\item{n.tips}{A single integer (>=1) or vector of integers specifying the 
number of tips in each branch. When 'n.tips' is a vector, its length 
should be equal to 'n.branches'.}
}
\value{
A \code{phylo} object.
}
\description{
The \code{symmetricBranching} function generates random binary 
trees with branches exhibiting various symmetry patterns, all mapping to 
the ancestral nodes of a reference tip. This generator is intended for 
assessing potential biases in null distributions.
}
\examples{
# Make a tree with 3 branches, 2 tips each, all branches
# anchored to the ancestral nodes of a single reference tip.
phy <- symmetricBranching(n.branches = 3 , n.tips = 2)
# plot(phy, type = "cladogram")

}
\seealso{
\code{\link{simulateRogs}}
}
\author{
Sysbiolab.
}
