% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Bridge-class}
\alias{Bridge-class}
\title{Bridge: An S4 class for Bridge analysis}
\value{
An S4 class object.
}
\description{
Bridge: An S4 class for Bridge analysis
}
\section{Slots}{

\describe{
\item{\code{cogids}}{Object of class \code{"data.frame"}, a data frame with COG data.}

\item{\code{tree}}{Object of class \code{"phylo"}, a given species tree.}

\item{\code{spbranches}}{Object of class \code{"data.frame"}, a data frame 
listing branches of a given species tree.}

\item{\code{orthoroot}}{Object of class \code{"data.frame"}, a data.frame with 
results from the 'runBridge' function.}

\item{\code{orthoct}}{Object of class \code{"data.frame"}, a data.frame with 
results from the 'Bridge' constructor.}

\item{\code{status}}{Object of class \code{"character"}, a string specifying the 
status of the Bridge object based on the available methods.}
}}

\section{Constructor}{

see \code{\link{newBridge}} constructor.
}

