% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{getBridge,Bridge-method}
\alias{getBridge,Bridge-method}
\alias{getBridge}
\title{Accessors for fetching slots from a Bridge object}
\usage{
\S4method{getBridge}{Bridge}(gbr, what = "status")
}
\arguments{
\item{gbr}{A preprocessed \linkS4class{Bridge} class object.}

\item{what}{A single character value specifying which information should 
be retrieved from the slots. 
Options: 'status', 'results', 'tree', 'ogids', 'orthoroots',
'branches', 'roots'.}
}
\value{
Content from slots in the \linkS4class{Bridge} object.
}
\description{
\code{getBridge} retrives information from
individual slots available in a Bridge object.
}
\examples{
# Load datasets used for demonstration
data(ogdata)

# Create an object of class 'Bridge' for H. sapiens
gbr <- newBridge(ogdata, phyloTree, refsp="9606", ogids=ogids)

# Get the 'status' slot in gbr
getBridge(gbr, what = 'status')

}
