% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{newBridge}
\alias{newBridge}
\title{Constructor of Bridge-class objects}
\usage{
newBridge(ogdata, phyloTree, refsp, ogids = NULL, verbose = TRUE)
}
\arguments{
\item{ogdata}{A data frame with COG data.}

\item{phyloTree}{Object of class \code{phylo}.}

\item{refsp}{A single string or integer specifying the reference species 
to be used in the Bridge algorithm. The \code{refsp} must be listed in 
the \code{tip.label} of the \code{phylo} object.}

\item{ogids}{An optional character vector listing which OGs should be
evaluated by the Bridge algorithm.}

\item{verbose}{A single logical value specifying to display detailed 
messages (when \code{verbose=TRUE}) or not (when \code{verbose=FALSE}).}
}
\value{
A preprocessed \linkS4class{Bridge} class object.
}
\description{
\code{Bridge} is a constructor of Bridge-class objects.
}
\examples{
# Load datasets used for demonstration
data(ogdata)

# Create an object of class 'Bridge' for H. sapiens
gbr <- newBridge(ogdata, phyloTree, refsp="9606", ogids=ogids)

}
\seealso{
\code{\link{runBridge}}
}
\author{
Sysbiolab.
}
