% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{runBridge,Bridge-method}
\alias{runBridge,Bridge-method}
\alias{runBridge}
\title{Run gene root inference with the Bridge algorithm}
\usage{
\S4method{runBridge}{Bridge}(gbr, penalty = 2, threshold = 0.5, verbose = TRUE)
}
\arguments{
\item{gbr}{A preprocessed \linkS4class{Bridge} class object.}

\item{penalty}{A single number (>0) specifying the penalty used in the 
rooting algorithm (see details).}

\item{threshold}{A single number in `[0,1)`, specifying a local threshold 
used in the Bridge algorithm (see details).}

\item{verbose}{A single logical value specifying to display detailed 
messages (when \code{verbose=TRUE}) or not (when \code{verbose=FALSE}).}
}
\value{
A preprocessed \linkS4class{Bridge} class object, including 
results from the Bridge algorithm.
}
\description{
The \code{runBridge} function assesses the 
evolutionary root of a gene based on the distribution of its orthologs.
}
\details{
The 'Bridge' algorithm addresses the problem of finding the evolutionary 
root of a feature in an phylogenetic tree. The method infers the probability
that such feature was present in the Last Common Ancestor (LCA) of a given 
lineage. Events like horizontal gene transfer, gene deletion, and de novo 
gene formation add noise to vertical heritage patterns. The 'runBridge' 
function assesses the presence and absence of orthologs in the extant
species of a phylogenetic tree in order to build a probability 
distribution, which is used to identify vertical heritage patterns.

The 'penalty' argument weighs gene gain and loss; penalty=1 indicates equal 
probability; penalty > 1 indicates higher probability of gene loss while 
penalty < 1 indicates higher probability of gene gain (penalty value should
be greater than zero; default \code{penalty = 2}).

After the probability distribution is calculated for a given lineage, 
then a rooting algorithm is used to search the LCA that provides the best  
vertical heritage pattern. The rooting algorithm finds the optimal point 
that splits the probability distribution into two components: one enriched
with the queried feature (supporting vertical heritage in the lineage)
and another with low evidence in favor of the feature's presence.
The 'threshold' sets the tolerance for the discrimination between the 
two components (default \code{threshold = 0.5}).
}
\examples{
# Load datasets used for demonstration
data(ogdata)

# Create an object of class 'Bridge' for H. sapiens
gbr <- newBridge(ogdata, phyloTree, refsp="9606", ogids=ogids)

# Run the Bridge algorithm
gbr <- runBridge(gbr)

}
\seealso{
\code{\link{newBridge}}
}
\author{
Sysbiolab.
}
